requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
deepmerge==2.0
idna==3.11
packaging==26.0
PyJWT==2.12.1
python-arango==8.3.0
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
tabulate==0.10.0
urllib3==2.6.3
